requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# YAML Configuration
PyYAML>=6.0

# Image Augmentation (optional but recommended)
albumentations>=1.3.0

# Metrics and Evaluation
scikit-learn>=1.3.0

# For YOLOv5 (if using detection stage)
# opencv-python>=4.8.0

# Jupyter (for notebooks)
# jupyter>=1.0.0
# ipykernel>=6.25.0
